Auth gran type changed from implicit to Authorization code in order to prepare the UserWorkManager to schedule automatically updates of the token in the background

+++ app/src/main/java/com/victorbg/racofib/view/ui/notes/NotesFragment.java
@@ -171,7 +171,7 @@ public class NotesFragment extends BaseFragment implements Observer<List<Note>>,
         publicationsViewModel.reload();
         swipeRefreshLayout.setRefreshing(true);
         publicationsViewModel.getPublications().observe(this, listResource -> {
-            Timber.d("Data observed with status %s and time %f", listResource.status.toString(), (float) System.currentTimeMillis());
+            Timber.d("Data observed with status %s and time %d", listResource.status.toString(), System.currentTimeMillis());
 
             new Handler().postDelayed(() -> {
                 /*
